requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
asn1crypto==1.5.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cloudscraper==1.2.60
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
Django==4.0.4
django-filter==22.1
django-htmx==1.12.1
django-import-export==2.8.0
django-snowflake==3.2a2
django-tables2==2.4.1
et-xmlfile==1.1.0
greenlet==1.1.2
idna==3.3
import-export==0.2.67.dev6
install==1.3.5
joblib==1.1.0
lml==0.1.0
loguru==0.6.0
lxml==4.8.0
MarkupPy==1.14
mssql-django==1.1.3
multiprocess==0.70.13
nltk==3.7
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
pandas==1.4.3
pathos==0.2.9
pox==0.3.1
ppft==1.7.6.5
pycparser==2.21
pycryptodomex==3.15.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
PyExecJS==1.5.1
PyJWT==2.4.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-toolbelt==0.9.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
snowflake-connector-python==2.7.11
snowflake-sqlalchemy==1.4.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
tablib==3.2.1
threadpoolctl==3.1.0
tqdm==4.64.0
translators==5.2.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
win32-setctime==1.1.0
xlrd==2.0.1
xlwt==1.3.0
